+++ resources/views/components/emails/footer-html.blade.php
@@ -9,3 +9,6 @@
 <p>
     @lang('ig-common::layouts.provider', ['link' => $providerLink, 'year' => date('Y')])
 </p>
+<p>
+    {{ $ip }} @if($timezone)({{ $timezone }})@endif
+</p>